cookbook/examples/workflows/B2BAgents/requirements.txt
agno
google-generativeai
requests
sqlalchemy
python-dotenv

